Moved two functions out of users-model and tracks-model to a helper file to make code a bit more DRY
