+++ tracks/tracks-model.js
@@ -7,7 +7,6 @@ module.exports = {
     getTracksByName,
     getTracksByArtistId,
     getClosestTracks,
-    mapTracks,
     getClosestValues
 };
 
@@ -58,12 +57,6 @@ function getTracksByArtistId(artist_id){
         .where({ artist_id });
 }
 
-async function getTrackByTrackId(track_id){
-    const track = await db('tracks').where(track_id).first();
-
-    return track;
-}
-
 async function getMeanValue(){
     const tracks = await db('tracks');
     
@@ -104,15 +97,4 @@ async function getClosestValues(target, page_number) {
     let result = Object.entries(json);
 
     return result;
-}
-
-async function mapTracks(array){
-    let promises = [];
-
-    for(let i =0; i< array.length; i++){
-        let track = getTrackByTrackId({track_id: array[i][0]});
-        promises.push(track);
-    }
-
-    return Promise.all(promises);
 }
