+++ users/users-model.js
@@ -7,8 +7,7 @@ module.exports = {
     addPostiveTrack,
     addNegativeTrack,
     getUserPredictedTracks,
-    getUserFitValues,
-    mapTracks
+    getUserFitValues
 };
 
 async function addUser(user){
@@ -127,21 +126,4 @@ async function getMeanValue(){
     let meanvalue = { mean_values: json };
 
     return meanvalue;
-}
-
-async function getTrackByTrackId(track_id){
-    const track = await db('tracks').where(track_id).first();
-
-    return track;
-}
-
-async function mapTracks(array){
-    let promises = [];
-
-    for(let i =0; i< array.length; i++){
-        let track = getTrackByTrackId({track_id: array[i][0]});
-        promises.push(track);
-    }
-
-    return Promise.all(promises);
 }
